+++ plugins/code_highlight.ts
@@ -66,6 +66,13 @@ export function codeHighlight(userOptions?: Options) {
   }
 
   return (site: Site) => {
+    if (site._data.codeHighlight) {
+      log.error(
+        `The plugin "${site._data.codeHighlight}" is already registered for the same purpose as "codeHighlight". Registering "codeHighlight" may lead to conflicts and unpredictable behavior.`,
+      );
+    }
+    site._data.codeHighlight = "codeHighlight";
+
     site.process([".html"], processCodeHighlight);
 
     if (options.theme) {